extract olm/Session into own file
